requirements.txt
influxdb
pyyaml
